Final modifications

+++ database/migrations/2018_12_16_004136_create_sessions_table.php
@@ -17,7 +17,7 @@ class CreateSessionsTable extends Migration
             $table->increments('id');
             $table->timestamps();
 
-            $table->integer('hours');
+            $table->float('hours', 4, 2);
         });
     }
